Lesson 5. Java 19 -> Java 11

+++ src/main/java/student_vladislav_romanov/lesson_5/level_2/Task_14.java
@@ -17,7 +17,7 @@ class Task_14 {
             arraySum += number;
         }
 
-        System.out.println(arraySum / numbers.length);
+        System.out.println(arraySum / (double) numbers.length);
     }
 
 }
